Rename executable package for clarity
