+++ config/webservice/route.go
@@ -1,4 +1,4 @@
-package myservice
+package webservice
 
 import (
 	"github.com/go-chi/chi"
